Create a Provider Composition Pattern
